Set workspace names both before and after reordering
The "Space Bar" extension now listens for when other extensions reorder
workspaces and attempts to reorder the names accordingly (see
<https://github.com/christopher-l/space-bar/commit/bbc7b07>). However,
this conflicts with this extension's reordering of names, which happens
prior to reodering the workspaces for compatibility with other
extensions (see `0243a21`). Thus, the best way to guarantee
compatibility with all extensions is to set the names both before and
after reordering workspaces.